fix: video generation
bug: camera is very hard to do, i dont know how to do it so ill let it there until i find a solution (or not).

+++ include/core/Core.hpp
@@ -45,8 +45,8 @@ namespace VC
     private:
 
         ///< Window size
-        const size_t _width;
-        const size_t _height;
+        const int _width;
+        const int _height;
 
         ///< Framerate of the video
         const size_t _framerate;
